requirements/production.txt
http://www.crummy.com/software/BeautifulSoup/unreleased/4.x/BeautifulSoup-4.0b.tar.gz
Django==1.3
django-grappelli==2.3.3
requests==0.6.1
wsgiref==0.1.2
mechanize==0.2.5
django-pagination==1.0.7
django-sentry==1.8.10
-e git://github.com/django-mptt/django-mptt.git@0af02a95877041b2fd6d458bd95413dc1666c321#egg=django-mptt
docutils==0.8.1
django-markitup==1.0.0
Creoleparser==0.7.4
Pillow==1.7.5
slate==0.3
ghostscript==0.4.1
-e git://github.com/rosarior/django-sendfile.git@f85a1737ebf6464c85c81d75dfeda03efe691b7e#egg=django-sendfile
django-compressor==1.1
